+++ app/move/config/index.js
@@ -1,13 +1,11 @@
 const assignConfig = require('./assign.config')
 const cancelConfig = require('./cancel.config')
-const reviewConfig = require('./review.config')
 const unassignConfig = require('./unassign.config')
 const updateConfig = require('./update.config')
 
 module.exports = {
   assignConfig,
   cancelConfig,
-  reviewConfig,
   unassignConfig,
   updateConfig,
 }
