+++ app/move/controllers/index.js
@@ -1,7 +1,6 @@
 const assign = require('./assign')
 const Cancel = require('./cancel')
 const confirmation = require('./confirmation')
-const Review = require('./review')
 const timeline = require('./timeline')
 const Unassign = require('./unassign')
 const update = require('./update')
@@ -11,7 +10,6 @@ module.exports = {
   assign,
   Cancel,
   confirmation,
-  Review,
   timeline,
   Unassign,
   update,
