+++ app/move/fields/index.js
@@ -35,7 +35,6 @@ const policeNationalComputerUpdate = require('./police-national-computer.update'
 const prisonRecallComments = require('./prison-recall-comments')
 const prisonTransferComments = require('./prison-transfer-comments')
 const prisonTransferType = require('./prison-transfer-type')
-const reviewFields = require('./review')
 const servingYouthSentence = require('./serving-youth-sentence')
 const shouldSaveCourtHearings = require('./should-save-court-hearings')
 const specialVehicleCheck = require('./special-vehicle-check')
@@ -136,6 +135,5 @@ module.exports = {
   assignFields,
   cancelFields,
   createFields,
-  reviewFields,
   updateFields,
 }
