+++ app/move/steps/index.js
@@ -1,13 +1,11 @@
 const assign = require('./assign')
 const cancel = require('./cancel')
-const review = require('./review')
 const unassign = require('./unassign')
 const update = require('./update')
 
 module.exports = {
   assign,
   cancel,
-  review,
   unassign,
   update,
 }
